added resource provider, connect to Api

+++ optimogroup-task/src/models/employee/Employee.ts
@@ -5,7 +5,7 @@ import { Resource } from "../resource";
 
 export class Employee extends Resource {
     name: string | any 
-    like: BigInt | any 
+    like: BigInteger | any 
     description: string | any 
     avatar: File[] | any 
     jobId: Job | any
